backend/requirements.txt
Flask==3.0.0
pytesseract==0.3.10
Pillow==10.1.0
gunicorn==21.2.0
requests==2.31.0
google-cloud-firestore==2.13.1
pytest==8.3.4
ruff==0.8.4